divided program into classes and added comment lines(not very good)
