+++ CardSign.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace cardGameSix
+{
+    public enum Sign : int { HEARTS = 1, DIAMONDS = 2, SPADES = 3, CLUBS = 4 };
+    // Creating a class for each sign of the playing card.
+}
