+++ CardValue.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace cardGameSix
+{
+    public enum Value : int { NINE = 9, TEN = 10, JACK = 12, QUEEN = 13, KING = 14, ACE = 15 }
+    // Creating a class for the numeric value of each playing card.
+}
